requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.3
pydantic==2.9.2
PyJWT==2.9.0

# Security tooling CLIs used by scanners
semgrep==1.99.0
bandit==1.7.9

